# Conflicts:
#	bench/bench-atlas-concurrency.py
#	crates/spark-model/src/layers/ops/dispatch_helpers.rs
#	crates/spark-server/src/scheduler/helpers.rs
#	scripts/realistic_soak.py

+++ kernels/strix-hip/common/prefill_paged_compute.cuh
@@ -230,7 +230,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
@@ -238,7 +238,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -326,7 +326,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -334,7 +334,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_V[k_off + k][d_col];
+                        bb[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
@@ -383,11 +383,15 @@ extern "C" __global__ void KERNEL_NAME(
 }
 
 // ==========================================================================
-// BR=64 entry (KERNEL_NAME##_64). On AMD/gfx1151 the BR=64 large-chunk paged
-// prefill kernels are COMPILE-ONLY (force_br32_prefill routes all dispatch to
-// the BR=32 kernel above — see HARDWARE.toml / paged_attn.rs). To keep LDS
-// within RDNA3.5's 64 KB cap and the grid math harmless, BR64 is clamped to 32
-// here (mirroring inferspark_prefill_wmma.cu). The wave32 WMMA + smem-softmax
+// BR=64 entry (KERNEL_NAME##_64). On AMD/gfx1151 these _64 paged prefill
+// kernels ARE still dispatched (paged_attn.rs picks them on chunk length
+// alone), so BR64 is clamped to 32 here both to keep LDS within RDNA3.5's
+// 64 KB cap and to make the tiling correct; the host grid is clamped to match
+// by cfg!(atlas_scale) in ops/prefill_attn_main_{a,b}.rs, and the two MUST
+// agree. (An earlier comment claimed a `force_br32_prefill` HARDWARE.toml key
+// routed dispatch away from them — no such routing existed, the key had no
+// reader, and it has been removed.) Mirrors inferspark_prefill_wmma.cu. The
+// wave32 WMMA + smem-softmax
 // body is identical to the BR=32 kernel; only the entry symbol differs so the
 // registry links. NVIDIA keeps BR64=64 verbatim.
 #define BR64 32
@@ -514,14 +518,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
                     unsigned int key_row = nt * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -597,14 +601,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
                     unsigned int d_col = (pv_n_start + nt) * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_V[k_off + k][d_col];
+                        bb[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
